Sprint Code Redone
I redid my code for the dropdown menu. Everything should now work as intended. Whoever did the pull request conflict resolution for the age input did not clean it well. There was a duplicate age field and a duplicate useState for age. I had to deal with an extra dependency causing an error. I uninstalled that package. A new package must be installed to work with the dropdown which can be installed from "react-native-picker-select". Just do "npm install reactnative-picker-select" to install it.
There is one issue I was unable to fix where the input field for Age keeps resetting to zero after the first age input being cleared. Causing the need for a page reload.